Title: 사각형 채우기 3, Time: 150ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240321/사각형 채우기 3/rectangle-fill-3.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n=sc.nextInt();
-        long[] dp = new int[n+1];
+        long[] dp = new long[n+1];
         int MOD = 1000000007;        
         dp[0]=1;
         dp[1]=2;
